refactoring vol 4
