+++ src/components/Map/behaviour/useUpdateStyle.tsx
@@ -13,7 +13,7 @@ import { makinaAfricaStyle } from '../styles/makinaAfricaStyle';
 import {
   CLIMBING_SPRITE,
   climbingLayers,
-} from '../styles/layers/climbingLayers';
+} from '../climbingTiles/climbingLayers';
 import { EMPTY_GEOJSON_SOURCE, OSMAPP_SPRITE } from '../consts';
 import { fetchCrags } from '../../../services/fetchCrags';
 import { intl } from '../../../services/intl';
@@ -22,7 +22,7 @@ import { setUpHover } from './featureHover';
 import { layersWithOsmId } from '../helpers';
 import { Theme } from '../../../helpers/theme';
 import { addIndoorEqual, removeIndoorEqual } from './indoor';
-import { addClimbingTilesSource } from './climbingTiles/climbingTilesSource';
+import { addClimbingTilesSource } from '../climbingTiles/climbingTilesSource';
 
 const ofrBasicStyle = {
   ...basicStyle,
@@ -64,28 +64,6 @@ const addRasterOverlay = (
   // TODO maxzoom 19 only for snow overlay
 };
 
-const addClimbingOverlay = (style: StyleSpecification, map: Map) => {
-  // style.sources.climbing = EMPTY_GEOJSON_SOURCE;
-  style.layers.push(
-    ...climbingLayers.map((x) => ({
-      ...x,
-      source: 'climbingTiles',
-      'source-layer': 'groups',
-    })),
-  ); // must be also in `layersWithOsmId` because of hover effect
-  style.sprite = [...OSMAPP_SPRITE, CLIMBING_SPRITE];
-
-  // fetchCrags().then(
-  //   (geojson) => {
-  //     const geojsonSource = map.getSource('climbing') as GeoJSONSource;
-  //     geojsonSource?.setData(geojson); // TODO can be undefined at first map render
-  //   },
-  //   (error) => {
-  //     console.warn('Climbing Layer failed to fetch.', error); // eslint-disable-line no-console
-  //   },
-  // );
-};
-
 const addOverlaysToStyle = (
   map: Map,
   style: StyleSpecification,
@@ -98,8 +76,6 @@ const addOverlaysToStyle = (
       switch (key) {
         case 'climbing':
           addClimbingTilesSource(style);
-
-          //addClimbingOverlay(style, map);
           break;
 
         case 'indoor':
